+++ test/entropies/estimators/lord.jl
@@ -13,8 +13,8 @@ npts = 20000 # a bit fewer points than for other tests, so tests don't take fore
 ea = entropy(Lord(k = 20), rand(npts))
 ea_n3 = entropy(Lord(k = 20, base = 3), randn(npts))
 
-@test U - max(0.03, U*0.03) ≤ ea ≤ U + max(0.03, U*0.03)
-@test N_base3 * 0.98 ≤ ea_n3 ≤ N_base3 * 1.02
+@test N_base3 * 0.96 ≤ ea_n3 ≤ N_base3 * 1.03
+@test U - max(0.05, U*0.03) ≤ ea ≤ U + max(0.03, U*0.03)
 
 x = rand(1000)
 @test_throws ArgumentError entropy(Renyi(q = 2), Lord(), x)