+++ test/utils/utils.jl
@@ -1,4 +1,4 @@
 @testset "Utility methods" begin
-    testfile("fasthist.jl")
+    #testfile("fasthist.jl")
     testfile("encoding.jl")
 end